requirements.txt
# 의존성 패키지 목록

flask==2.0.1
flask-cors==3.0.10
opencv-python==4.5.4.60
numpy==1.21.4
insightface==0.6.2
langchain==0.1.0
langchain-community==0.0.13
faiss-cpu==1.7.3
google-generativeai>=0.3.1
python-dotenv==1.0.0